hooked up to wifi and sending data
