--- on-device/pymakr.conf
@@ -1,10 +1,0 @@
-{
-  "py_ignore": [
-    ".vscode",
-    ".gitignore",
-    ".git",
-    "env",
-    "venv"
-  ],
-  "name": "on-device"
-}
